requirements.txt
discord
ipython
pytest
python-dotenv
replit
transitions
python-slugify
